+++ product-service/database/products.ts
@@ -1,41 +1,35 @@
 export const products = [
   {
-    id: 1,
     count: 4,
     description: 'Short Product Description1',
     price: 2.4,
-    title: 'New Product1'
+    title: 'Product1'
   },
   {
-    id: 2,
     count: 1,
     description: 'Short Product Description2',
     price: 21.4,
     title: 'Brand New Product!'
   },
   {
-    id: 3,
     count: 1,
     description: 'Short Product Description3',
     price: 21.4,
     title: 'New Product3'
   },
   {
-    id: 4,
     count: 1,
     description: 'Short Product Description4',
     price: 41.4,
     title: 'New Product4'
   },
   {
-    id: 5,
     count: 1,
     description: 'Short Product Description5',
     price: 51.4,
     title: 'New Product5'
   },
   {
-    id: 6,
     count: 1,
     description: 'Short Product Description6',
     price: 61.4,
